# Re-organize hash table access in dict page for performance