```
name                        old time/op    new time/op    delta
WriteCSVPlainDictionary-44    14.0ms ± 4%    11.2ms ± 3%  -19.82%  (p=0.000 n=19+20)

name                        old alloc/op   new alloc/op   delta
WriteCSVPlainDictionary-44    3.17MB ± 0%    3.17MB ± 0%   -0.00%  (p=0.020 n=19+20)

name                        old allocs/op  new allocs/op  delta
WriteCSVPlainDictionary-44     21.8k ± 0%     21.8k ± 0%   -0.00%  (p=0.004 n=20+20)
```